+++ Sources/Alchemy/Core/HTTP/HTTPRouterResponder.swift
@@ -11,19 +11,21 @@ struct HTTPRouterResponder: HTTPResponder {
     func respond(to request: HTTPRequest) -> EventLoopFuture<HTTPResponse> {
         Router.globalMiddlewares
             // First apply global middlewares, in order.
-            .reduce(Loop.future(value: request)) {
+            .reduce(EventLoopFuture<HTTPRequest>.new(request)) {
                 $0.flatMap($1.intercept)
             }
             // Then, send to the router.
             .flatMap { request in
                 guard let response = self.router.handle(request: request) else {
                     // If the router doesn't handle the request, return a 404.
-                    return Loop.future(value: HTTPResponse(status: .notFound, body: nil))
+                    return .new(HTTPResponse(status: .notFound, body: nil))
                 }
                 
                 // If the router CAN handle the request, turn the returned
                 // `HTTPResponseEncodable` into an `HTTPResponse`.
-                return response.throwingFlatMap { try $0.encode() }
+                return response.flatMap { res in
+                    catchError { try res.encode() }
+                }
             }
     }
 }
