+++ Sources/Alchemy/Utilities/Array+Utilities.swift
@@ -14,6 +14,7 @@ extension Array where Element: Hashable {
 }
 
 extension Array {
+    /// Allows for safe array lookup, returning nil of the index is out of bounds.
     public subscript(safe index: Int) -> Element? {
         guard index >= 0, index < endIndex else {
             return nil
