+++ Sources/Alchemy/Utilities/Metatype+Utilities.swift
@@ -1,7 +1,15 @@
+/// Gives a unique identifier for a metatype.
+///
+/// - Parameter metatype: the type to provide a unique identifier for.
+/// - Returns: a unique identifier for this type.
 public func identifier<T>(of metatype: T.Type) -> ObjectIdentifier {
     ObjectIdentifier(metatype)
 }
 
+/// Returns the string name of this type.
+///
+/// - Parameter metatype: the type to name.
+/// - Returns: the name of the type.
 public func name<T>(of metatype: T.Type) -> String {
     "\(metatype)"
 }
